CI: Add FreeBSD builds through Cirrus CI.
When the github application Cirrus CI (https://cirrus-ci.org/) is
installed, this build description file adds FreeBSD build checks to
changes and PRs in github.

There are two build tasks, with all options enabled on two different
FreeBSD versions. Currently jack2 is built from the develop branch as a
dependency, this can be changed once there is an official package for
1.9.20 is available.