macOS support (#72)
* Fix build assumptions for librt and opus

meson.build:
Only require librt when building zalsa. The library is also not
available on macOS, which makes it even more so important to not search
for it and require it by default.
Require the existence of `opus/opus_custom.h` to be able to build with
opus support. The existence of the header file depends on whether opus
has been built using the `--enable-custom-modes` configure option.

* Force the creation of symlinks

scripts/meson_create_symlink:
Force the creation of symlinks so that existing files are overwritten.

* Add build workflow for latest macOS

.github/workflows/build.yml:
Extend build workflow by job for building on latest macOS.